// client/src/js/bindings/BedrockStatus.ts
// This file was generated by [ts-rs](https://github.com/Alaydriem/ts-rs). Do not edit this file manually.

export interface BedrockStatus {
    proxy_running: boolean;
    realms_running: boolean;
    xbox_authenticated: boolean;
    proxy_target_host: string | null;
    proxy_target_port: number | null;
    proxy_listen_port: number | null;
    active_realm_id: number | null;
    active_realm_name: string | null;
}